Split up into multiple files; update docs
