+++ services/telemetry/README.md
@@ -4,14 +4,12 @@ Service that serves plane telemetry and mission data over a REST API.
 
 ## Known Problems
 
-If the download mission action from dronekit times out, then new mission data
-cannot be fetched. Unfortunately, this can happen quite often.
+It is not safe to transact with the plane concurrently with other applications.
+This should be the only application directly writing to the plane.
 
 ## Running the Image
 
 The `CXN_STR` environment variable is used for connecting to the plane.
-Optionally, the `CXN_TIMEOUT` and `CXN_RETRY` environment variables can set the
-connection timeout time and whether or not to retry connecting on startup.
 
 ```
 $ docker run -it -p 5000:5000 \
@@ -23,6 +21,10 @@ $ docker run -it -p 5000:5000 \
 
 Note that all the Protobuf endpoints can also return JSON as well if the Accept
 header is set to `application/json`.
+
+All requests long-poll until a connection to the plane has been established.
+If an error occurs, a `504` status code is returned, or `503` if not enough
+information has been captured from the plane at request time.
 
 - `GET /api/alive`
 
@@ -47,6 +49,25 @@ header is set to `application/json`.
   If any of the plane's lat, lon, alt, yaw, roll, pitch aren't loaded, an empty
   response with a `204` status code will be returned.
 
+- `GET /api/overview`
+
+  An overview of all of the plane's last known values.
+
+  On successful response: `200` status code with `telemetry::Overview`
+  Protobuf message.
+
+  If information currently unavailable: `503` status code with an empty body.
+
+- `GET /api/mission`
+
+  JSON mission data from the plane parsed by the MAVLink library.
+
+  Very similar to `/api/raw-mission`.
+
+  On successful response: `200` status code with JSON data.
+  On failure: `504` status code with a JSON field called `err` stating
+  an error reason.
+
 - `GET /api/raw-mission`
 
   Raw mission data from the plane.
@@ -55,3 +76,11 @@ header is set to `application/json`.
 
   On successful response: `200` status code with `telemetry::RawMission`
   Protobuf message.
+
+- `POST /api/raw-mission`
+
+  Pushes raw mission data to the plane, overwriting the previous mission.
+
+  The request body must be a `telemetry::RawMission` Protobuf message.
+
+  On successful response: `200` status code with an empty body.
